day 1: full solution
The naive solution trying to simply replace text-digits with their
numeric counterpart had the problem of overlaps leading to the loss of
the digit processed last of the overlapping ones.

Instead we find all the digit-like parts of the text with a regex and
keep track of its index. We can then find the minimum and maximum
indecies in the list to find our output sum